Calculate Market Cap

+++ app/src/main/java/com/lukaslechner/coroutineusecasesonandroid/usecases/flow/usecaseDebounce/database/CryptoCurrencyEntity.kt
@@ -8,7 +8,7 @@ import com.lukaslechner.coroutineusecasesonandroid.usecases.flow.mock.CryptoCurr
 data class CryptoCurrencyEntity(
     @PrimaryKey val symbol: String,
     val name: String,
-    val totalSupply: Int,
+    val totalSupply: Float,
     val priceUsd: Float
 )
